Add policy for /usr/libexec/samba/samba-bgqd
samba-bgqd is a daemon enumerating cups printers and storing the
gathered information in a cache. Since gathering the information about
one printer takes about 1 sec, in environments with many printers
the process may get killed before finishing. Providing a way to start
samba-bgqd as a systemd service and always having it running in the
background is a solution to this issue.

Resolves: RHEL-53124