Allow samba-dcerpcd work with ctdb cluster
ctdb is Clustered Database based on Samba's Trivial Database (TDB).
In a clustered Samba setup, /usr/libexec/samba/samba-dcerpcd,
when invoked on demand to list the services available from a server,
needs to be allowed to connect to the ctdbd service over a unix socket
and map its database files.

Resolves: RHEL-14759