Allow svirt_t the sys_rawio capability
This permission is needed when a vm needs raw access to SCSI disks.

libvirt explicitly grants the qemu process CAP_SYS_RAWIO only if it has
a disk with the property enabled in the definition and otherwise a qemu
process started by libvirt will not have that capability.

Note the previous commit 181dc5c40f52 ("Allow svirt_t the sys_rawio
capability") allowed the permission actually to the svirt_tcg_t domain.

Resolves: RHEL-61235