allow kdm to create /root/.kde/ with correct label
When the kdm service is started, it wants to create the .kde directory
under /root/, but SELinux denies that action. When the /root/.kde
directory exists, the kdm service wants to create a symlink in it,
but SELinux denies that action too. The intended symlink should
point this way:
 * /root/.kde/cache-machine-<FQDN> --> /var/tmp/kdecache-root

SELinux policy should now label the newly created /root/.kde
directory correctly as config_home_t, which is consistent with
other definitions. Allow rule for creating a symlink is already
defined.

Resolves: bz#2275868