Allow smbd to load libcephfs proxy conditionally
The design for libcephfs proxy[1] was accepted sometime back with the
corresponding implementation subsequently merged[2] into upstream Ceph.
This additional way to integrate CephFS with Samba is comprised of two
components. The daemon i.e, libcephfsd is responsible for creating the
unix domain socket as a starting point of communication. On the other
side proxy library, which is designed[3] to be loaded within smbd,
connects to the already created socket and thereby communicates with
the daemon.

During the development and testing stages we happened to notice some AVC
denial messages in the audit logs:

type=AVC msg=audit(1738236295.070:1470): avc: denied { write } for pid=41822
  comm="smbd[192.168.12" name="libcephfsd.sock" dev="tmpfs" ino=1908
  scontext=system_u:system_r:smbd_t:s0 tcontext=unconfined_u:object_r:var_run_t:s0
  tclass=sock_file permissive=1

type=AVC msg=audit(1738236295.070:1470): avc: denied { connectto } for pid=41822
  comm="smbd[192.168.12" path="/run/libcephfsd.sock"
  scontext=system_u:system_r:smbd_t:s0 tcontext=unconfined_u:unconfined_r:unconfined_t:s0-s0:c0.c1023
  tclass=unix_stream_socket permissive=1

Therefore we propose a new boolean - samba_load_libcephfs_proxy, off by
default, to contain the necessary allow rules for smbd_t to operate in a
proxy mode using libcephfs.

[1] https://docs.ceph.com/en/latest/dev/libcephfs_proxy/
[2] https://github.com/ceph/ceph/pull/58376
[3] https://gitlab.com/samba-team/samba/-/merge_requests/3792

Signed-off-by: Anoop C S <anoopcs@cryptolab.net>