Rename winbind_rpcd_* types to samba_dcerpcd_*
In the current policy for samba, there are samba-dcerpcd and rpcd-*
processes labeled with winbind_rpcd_t. However, samba-dcerpcd is a
standalone DCERPC server, started by smbd or winbind. So to avoid
confusion that it is winbind only, all related types were renamed
to match the binary, in particular
- winbind_rpcd_t to samba_dcerpcd_t
- winbind_rpcd_exec_t to samba_dcerpcd_exec_t
- winbind_rpcd_var_run_t to samba_dcerpcd_var_run_t
- winbind_rpcd_tmp_t to samba_dcerpcd_tmp_t

The previous type names are kept as aliases for compatibility reasons.

Resolves: RHEL-14759